ml-pipeline/pyproject.toml
[project]
name = "hsi-ml-pipeline"
version = "0.1.0"
description = "ML pipeline for human-scene interaction: JOSH reconstruction + SAM3 segmentation + CLIP labeling"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "trimesh>=4.0.0",
    "open3d>=0.18.0",
    "smplx>=0.1.28",
    "roma>=1.3.0",
    "huggingface-hub>=0.20.0",
    "transformers>=4.36.0",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
    "tqdm>=4.66.0",
    "plyfile>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
